requirements/requirements.txt
# web
fastapi-jwt-auth==0.5.0
fastapi==0.63.0
fastapi_utils==0.2.1
httpx==0.17.0
python-multipart==0.0.5
starlette==0.13.6  # requires by fastapi==0.63.0
uvicorn==0.13.4

# databases
SQLAlchemy==1.3.23
aiopg==1.1.0
alembic==1.5.5
asyncpg==0.22.0
databases[postgresql]==0.4.1
psycopg2-binary==2.8.6
psycopg2==2.8.6

# asyncio utils
aiomisc==12.1.0
asynctempfile==0.5.0

# utils
bcrypt==3.2.0
filetype==1.0.7
linkpreview==0.1.12
minio==7.0.2
pydantic==1.8
python-dotenv==0.15.0
ujson==4.0.2
